requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3
ttkbootstrap==1.12.0
urllib3==2.4.0
